training/requirements.txt
# IntelFactor Training Pipeline — Python dependencies
# Install on Azure GPU VM: pip install -r training/requirements.txt
#
# NOTE: torch/torchvision are for training only (on x86 GPU VM).
#       Do NOT install PyTorch on Jetson — inference uses TensorRT.

# Core training
ultralytics>=8.2.0
torch>=2.2.0
torchvision>=0.17.0

# Dataset acquisition
roboflow>=1.1.0

# ONNX export and validation
onnx>=1.16.0
onnxruntime-gpu>=1.18.0    # for ONNX validation on GPU VM (not needed on Jetson)
onnxsim>=0.4.36            # ONNX simplifier (used by export_model.py)

# Augmentations
albumentations>=1.4.0

# Data science / analysis
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# AWS artifact upload (Phase 7)
boto3>=1.34.0

# Config / utilities
pyyaml>=6.0.1
